pc - refactor package names for startup and exclude from both jacoco and pitest
