+++ pom.xml
@@ -213,7 +213,7 @@
                         <exclude>**/edu/ucsb/cs156/courses/services/CurrentUserServiceImpl.*</exclude>
                         <exclude>**/edu/ucsb/cs156/courses/CoursesApplication.*</exclude>
                         <exclude>**/edu/ucsb/cs156/courses/CoursesApplicationRunner*.*</exclude>
-                        <exclude>**/edu/ucsb/cs156/courses/CoursesStartup.*</exclude>
+                        <exclude>**/edu/ucsb/cs156/courses/startup/*</exclude>
                         <exclude>**/edu/ucsb/cs156/courses/jobs/ScheduledJobs.*</exclude>
                     </excludes>
                 </configuration>
@@ -259,6 +259,10 @@
                         <param>edu.ucsb.cs156.courses.config.SpringFoxConfig</param>
                         <param>edu.ucsb.cs156.courses.config.MongoConfig</param>
                         <param>edu.ucsb.cs156.courses.config.MongoDevConfig</param>
+                        <param>edu.ucsb.cs156.courses.startup.CoursesStartup</param>
+                        <param>edu.ucsb.cs156.courses.startup.CoursesApplicationRunnerDevelopment</param>
+                        <param>edu.ucsb.cs156.courses.startup.CoursesApplicationRunnerProduction</param>
+                        <param>edu.ucsb.cs156.courses.jobs.ScheduledJobs</param>
                     </excludedClasses>
                     <excludedTestClasses>
                         <param>edu.ucsb.cs156.example.integration.*</param>
